FixedPointToolkit.Updates

FixedPointToolkit.Updates.BroydenMixing — Method

BroydenMixing(VIn::N, VOut::N, F::T ; F_args::Tuple = (), F_kwargs::Dict = Dict() , alpha::Float64 = 0.5, B::Z = alpha*Matrix(1.0I, length(VIn), length(VIn)), _extra...) --> Dict where {T<:Function, N<:Union{Number, Vector{<:Number}}, Z<:Union{Number,Matrix{<:Number}}}

Returns a dictionary containing the following

"VInNext" : simple mixing of VIn and VOut with weight (1-beta) and beta respectively, where beta = Scheduler(alpha), which may change with iteration.

"VOutNext" : The function F evaluated at VInNExt, with fixed args F_args and kwargs F_kwargs.

"Delta" : The norm difference b/w the next input and output.

"kwargs" : the kwargs of this function itself, to be used by a subsequent function call in the next fixed point iteration. Here, the B-matrix is updated every iteration according to Broyden mixing rules.

FixedPointToolkit.Updates.ScheduledMixing — Method

ScheduledMixing(VIn::N, VOut::N, F::T ; F_args::Tuple = (), F_kwargs::Dict = Dict() , alpha::Float64 = 0.5, iter::Int64 = 1, max_iter::Int64 = 1, Scheduler :: R = QuadraticScheduling, _extra...) --> Dict   where {T<:Function, N<:Union{Number, Vector{<:Number}}, R<:Function}

Returns a dictionary containing the following

"VInNext" : Simple mixing of VIn and VOut with weight (1-beta) and beta respectively, where beta = Scheduler(alpha), which may change with iteration.

"VOutNext" : The function F evaluated at VInNExt, with fixed args F_args and kwargs F_kwargs.

"Delta" : The norm difference b/w the next input and output.

"kwargs" : the kwargs of this function itself, to be used by a subsequent function call in the next fixed point iteration. Here, the alpha value is updated every iteration using a Scheduler.

FixedPointToolkit.Updates.SimpleMixing — Method

SimpleMixing(VIn::N, VOut::N, F::T ; F_args::Tuple = (), F_kwargs::Dict = Dict() , alpha::Float64 = 0.5, _extra...) --> Dict   where {T<:Function, N<:Union{Number, Vector{<:Number}}}

Returns a dictionary containing the following

"VInNext" : Simple mixing of VIn and VOut with weight (1-alpha) and alpha respectively.

"VOutNext" : The function F evaluated at VInNExt, with fixed args F_args and kwargs F_kwargs.

"Delta" : The norm difference b/w the next input and output.

"kwargs" : the kwargs of this function itself, to be used by a subsequent function call in the next fixed point iteration. Since this is simple mixing, alpha is held constant.

Hot to write your own Update

If you want to design your own update function, you need to follow certain guidelines

Define your function as follows (F is the function whose fixed point is being calculated)

function CustomUpdate(VIn::N, VOut::N, F::T ; F_args::Tuple = (), F_kwargs::Dict = Dict(), kwargs..., _extra...) :: Dict where {T<:Function, N<:Union{Number, Vector{<:Number}}}
    ##### Get VInNext by doing something on VIn, VOut, and stuff in kwargs.
    VOutNext    =   F(VInNext, F_args... ; F_kwargs...)

    ##### If you want to include scheduling, you can now include a function which updates the kwargs themselves --> call then new_kwargs (MUST contain all the keys of kwargs, even if they are unchanged)

    return Dict("VInNext"   => VInNext, 
                "VOutNext"  => VOutNext,
                "Delta"     => norm(VOutNext - VInNext) / sqrt(length(VInNext)), 
                "kwargs"    => new_kwargs)
end

Every other argument you pass to your CustomUpdate must be a keyword argument.

Unlike CustomScheduler, you will be able to pass whatever kwargs you want to CustomUpdate when you begin the Fixed Point iterations..